Switch database to SQL Server and fetch connectionstring
