+++ src/Infrastructure/EFUserRepositoryFactory.cs
@@ -1,4 +1,6 @@
-﻿using Infrastructure.Users;
+﻿using System.Diagnostics.CodeAnalysis;
+
+using Infrastructure.Users;
 
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Design;
@@ -9,14 +11,20 @@ namespace Infrastructure;
 /// <summary>
 /// Used at design-time by Entity Framework Core to generate the database migrations.
 /// </summary>
-public class PostgresUserRepositoryFactory : IDesignTimeDbContextFactory<EFUserRepository>
+[ExcludeFromCodeCoverage]
+public class EFUserRepositoryFactory : IDesignTimeDbContextFactory<EFUserRepository>
 {
     /// <inheritdoc />
     public EFUserRepository CreateDbContext(string[] args)
     {
         var optionsBuilder = new DbContextOptionsBuilder<EFUserRepository>();
-        _ = optionsBuilder.UseNpgsql(
-            "User ID=postgres;Password=password;Host=localhost;Port=5432;Database=contract_portal;");
+        string? dbConnectionstring = Environment.GetEnvironmentVariable(EnvironmentVariableKeys.DbConnectionString);
+        if (dbConnectionstring == null)
+        {
+            throw new ArgumentException("No environment variable defined for DbConnectionString");
+        }
+
+        _ = optionsBuilder.UseSqlServer(dbConnectionstring);
         return new EFUserRepository(
             optionsBuilder.Options,
             new LoggerFactory().CreateLogger<EFUserRepository>());
