+++ src/Infrastructure/EnvironmentVariableKeys.cs
@@ -1,0 +1,12 @@
+﻿namespace Infrastructure;
+
+/// <summary>
+/// All environment variable keys.
+/// </summary>
+public static class EnvironmentVariableKeys
+{
+    /// <summary>
+    /// The name of environment variable.
+    /// </summary>
+    public const string DbConnectionString = "prodigo_portal_db_connectionstring";
+}
